# docs/api/openapi/python-api.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nPython backend for Smart Ration HSD2C (side-by-side migration from the C#/.NET API).\n\n* Versioned: every route below answers under `/api/v1/...` as well as `/api/...` (same behaviour).\n* Every response uses the envelope `{success, message, data, errors}`; failures add `errorCode`.\n* `/api/*` routes that are not yet migrated are transparently served by the C# API\n  (fallback proxy); those do not appear in this document yet.\n",
    "title": "Smart Ration HSD2C API (Python)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/beneficiaries/{beneficiary_id}/collections": {
      "get": {
        "operationId": "beneficiary_collections_api_beneficiaries__beneficiary_id__collections_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary's collection history",
        "tags": [
          "people"
        ]
      }
    },
    "/api/beneficiaries/{beneficiary_id}/entitlement": {
      "get": {
        "operationId": "beneficiary_entitlement_api_beneficiaries__beneficiary_id__entitlement_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary's monthly entitlement",
        "tags": [
          "people"
        ]
      }
    },
    "/api/beneficiaries/{beneficiary_id}/family": {
      "get": {
        "operationId": "beneficiary_family_api_beneficiaries__beneficiary_id__family_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary's family",
        "tags": [
          "people"
        ]
      }
    },
    "/api/beneficiaries/{beneficiary_id}/full-profile": {
      "get": {
        "operationId": "beneficiary_full_profile_api_beneficiaries__beneficiary_id__full_profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary 360°: everything one screen needs",
        "tags": [
          "people"
        ]
      }
    },
    "/api/beneficiaries/{beneficiary_id}/verification": {
      "get": {
        "operationId": "beneficiary_verification_api_beneficiaries__beneficiary_id__verification_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary verification profile",
        "tags": [
          "people"
        ]
      }
    },
...
    "/api/families/{family_id}": {
      "get": {
        "operationId": "get_family_api_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "title": "Family Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A family",
        "tags": [
          "people"
        ]
      }
    },
    "/api/families/{family_id}/entitlement": {
      "get": {
        "operationId": "get_family_entitlement_api_families__family_id__entitlement_get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "title": "Family Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A family's monthly entitlement",
        "tags": [
          "people"
        ]
      }
    },
...
    "/api/public/beneficiaries/{reference}": {
      "get": {
        "operationId": "public_badge_api_public_beneficiaries__reference__get",
        "parameters": [
          {
            "in": "path",
            "name": "reference",
            "required": true,
            "schema": {
              "title": "Reference",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public verification badge (no login; non-sensitive fields only)",
        "tags": [
          "people"
        ]
      }
    },
...
    "/api/ration/collection/history/{beneficiary_id}": {
      "get": {
        "operationId": "collection_history_api_ration_collection_history__beneficiary_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Beneficiary's collection history (alias)",
        "tags": [
          "people"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard search (results scoped by role)",
        "tags": [
          "people"
        ]
      }
    },
...
  }
}